+++ package/Modules/02-Runtime/KIModuleRuntime.psm1
@@ -1,4 +1,4 @@
-﻿Set-StrictMode -Version Latest
+Set-StrictMode -Version Latest
 
 function Update-KIProcessPath {
     $machinePath = [Environment]::GetEnvironmentVariable('Path', 'Machine')