+++ package/Modules/06-Applications/KIModuleApplications.psm1
@@ -1,4 +1,4 @@
-﻿Set-StrictMode -Version Latest
+Set-StrictMode -Version Latest
 
 function Test-KIModuleApplications {
     param([Parameter(Mandatory)][object]$Context)