+++ package/Modules/07-Integration/KIModuleIntegration.psm1
@@ -1,4 +1,4 @@
-﻿Set-StrictMode -Version Latest
+Set-StrictMode -Version Latest
 
 function Test-KIModuleIntegration {
     param([Parameter(Mandatory)][object]$Context)